# back-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Juegos",
        "description": "API para la gestión de juegos, rondas y participantes. Permite crear juegos, gestionar rondas y administrar participantes con sus respectivos resultados.",
        "contact": {
            "name": "Soporte API",
            "email": "fabriciocervantesmendoza@gmail.com"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://opensource.org/licenses/AGPL-3.0"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor API"
        }
    ],
    "paths": {
...
        "/api/rondas": {
            "post": {
                "tags": [
                    "Rondas"
                ],
                "summary": "Crear una nueva ronda",
                "operationId": "ef201efdaade6bb1387bd983139300ff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "game_id",
                                    "participantes"
                                ],
                                "properties": {
                                    "game_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "participantes": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3,
                                            4
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ronda creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ronda iniciada exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Ronda"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/rondas/{ronda}/add-players": {
            "post": {
                "tags": [
                    "Rondas"
                ],
                "summary": "Agregar participantes a una ronda",
                "operationId": "165361aadfea254392c71b8120f3f97b",
                "parameters": [
                    {
                        "name": "ronda",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_ids"
                                ],
                                "properties": {
                                    "user_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            5,
                                            7,
                                            9
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Participantes agregados exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "3 participante(s) agregado(s) exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "participantes_agregados": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Participante"
                                                    }
                                                },
                                                "usuarios_duplicados": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación o ronda finalizada"
                    }
                }
            }
        },
        "/api/rondas/{ronda}/remove-player": {
            "delete": {
                "tags": [
                    "Rondas"
                ],
                "summary": "Remover un participante de la ronda",
                "operationId": "2553cd4fa6f40c62c2ac66fd760c92b7",
                "parameters": [
                    {
                        "name": "ronda",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 7
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Participante removido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Participante Juan Pérez removido exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Participante no encontrado"
                    },
                    "400": {
                        "description": "No se puede remover al ganador o ronda finalizada"
                    }
                }
            }
        },
        "/api/rondas/{ronda}/set-winner": {
            "patch": {
                "tags": [
                    "Rondas"
                ],
                "summary": "Establecer o quitar ganador de la ronda",
                "operationId": "ce919fde6e96b7202b00dc2aa23b14e9",
                "parameters": [
                    {
                        "name": "ronda",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "win"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "win": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ganador establecido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ganador establecido exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "participante": {
                                                    "$ref": "#/components/schemas/Participante"
                                                },
                                                "ronda_finalizada": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Participante no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "title": "Usuario",
                "description": "Modelo de usuario/jugador",
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "description": "ID único del usuario",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre completo del usuario",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "description": "Correo electrónico del usuario",
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "juan.perez@example.com"
                    },
                    "email_verified_at": {
                        "description": "Fecha de verificación del email",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Fecha de creación del usuario",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-10T08:15:30Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-20T14:22:45Z"
                    }
                },
                "type": "object"
            },
...
            "Ronda": {
                "title": "Ronda",
                "description": "Modelo de ronda de juego",
                "required": [
                    "id",
                    "fec",
                    "hora_ini",
                    "game_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID único de la ronda",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "fec": {
                        "description": "Fecha de la ronda",
                        "type": "string",
                        "format": "date",
                        "example": "2024-08-05"
                    },
                    "hora_ini": {
                        "description": "Hora de inicio de la ronda",
                        "type": "string",
                        "format": "time",
                        "example": "14:30:00"
                    },
                    "hora_fin": {
                        "description": "Hora de finalización de la ronda",
                        "type": "string",
                        "format": "time",
                        "example": "16:45:00",
                        "nullable": true
                    },
                    "game_id": {
                        "description": "ID del juego al que pertenece la ronda",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "game": {
                        "$ref": "#/components/schemas/Game"
                    },
                    "participantes": {
                        "description": "Lista de participantes en la ronda",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Participante"
                        }
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-05T14:30:00Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-05T16:45:00Z"
                    }
                },
                "type": "object"
            },
            "Participante": {
                "title": "Participante",
                "description": "Modelo de participante en una ronda",
                "required": [
                    "id",
                    "user_id",
                    "ronda_id",
                    "winner"
                ],
                "properties": {
                    "id": {
                        "description": "ID único del participante",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "description": "ID del usuario participante",
                        "type": "integer",
                        "format": "int64",
                        "example": 3
                    },
                    "ronda_id": {
                        "description": "ID de la ronda",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "winner": {
                        "description": "Indica si el participante es el ganador",
                        "type": "boolean",
                        "example": false
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "ronda": {
                        "$ref": "#/components/schemas/Ronda"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-05T14:30:00Z"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-05T16:45:00Z"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token JWT en el formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}